+++ rules/recording.go
@@ -82,9 +82,9 @@ func (rule *RecordingRule) Labels() labels.Labels {
 }
 
 // Eval evaluates the rule and then overrides the metric names and labels accordingly.
-func (rule *RecordingRule) Eval(ctx context.Context, queryOffset time.Duration, ts time.Time, query QueryFunc, _ *url.URL, limit int) (promql.Vector, error) {
+func (rule *RecordingRule) Eval(ctx context.Context, queryOffset, evalDelay time.Duration, ts time.Time, query QueryFunc, _ *url.URL, limit int) (promql.Vector, error) {
 	ctx = NewOriginContext(ctx, NewRuleDetail(rule))
-	vector, err := query(ctx, rule.vector.String(), ts.Add(-queryOffset))
+	vector, err := query(ctx, rule.vector.String(), ts.Add(-queryOffset).Add(-evalDelay))
 	if err != nil {
 		return nil, err
 	}